(prototype): Move login/out into navbar
