+++ frontend/src/components/login.rs
@@ -2,8 +2,13 @@
 
 use std::rc::Rc;
 
+use linkdoku_common::LoginFlowStart;
+use reqwest::StatusCode;
 use yew::prelude::*;
 use yew::Reducible;
+use yew_router::prelude::*;
+
+use crate::Route;
 
 use super::core::use_api_url;
 
@@ -77,7 +82,63 @@ pub fn login_user_provider(props: &UserProviderProps) -> Html {
     }
 }
 
-#[function_component(LogInOutButton)]
-pub fn login_inout_button() -> Html {
-    todo!()
+#[function_component(LoginButton)]
+pub fn login_button() -> Html {
+    let history = use_history().unwrap();
+    let start_google = use_api_url("/login/start/google");
+    let login_click = Callback::from(move |_| {
+        // User clicked login, so we need to redirect the user to the login flow
+        // startup
+        let history = history.clone();
+        let start_google = start_google.clone();
+        wasm_bindgen_futures::spawn_local(async move {
+            let response = reqwest::get(start_google)
+                .await
+                .expect("Unable to make call");
+            if response.status() != StatusCode::OK {
+                // some kind of error
+            } else {
+                let res: LoginFlowStart = response.json().await.unwrap();
+                match res {
+                    LoginFlowStart::Idle => history.push(Route::Root),
+                    LoginFlowStart::Redirect(url) => {
+                        gloo::utils::window().location().set_href(&url).unwrap();
+                    }
+                    LoginFlowStart::Error(err) => {
+                        gloo::console::log!("Failure doing login? {}", err);
+                        history.push(Route::Root);
+                    }
+                }
+            }
+        });
+    });
+
+    html! {
+        <button class={"button is-primary"} onclick={login_click}>
+            {"Login with Google"}
+        </button>
+    }
+}
+
+#[function_component(LogoutButton)]
+pub fn logout_button() -> Html {
+    let login_status_dispatch =
+        use_context::<LoginStatusDispatcher>().expect("Cannot get login status dispatcher");
+    let history = use_history().unwrap();
+    let clear_login = use_api_url("/login/clear");
+    let logout_click = Callback::from(move |_| {
+        let history = history.clone();
+        let login_status_dispatch = login_status_dispatch.clone();
+        let clear_login = clear_login.clone();
+        wasm_bindgen_futures::spawn_local(async move {
+            reqwest::get(clear_login).await.unwrap();
+            history.push(Route::Root);
+            login_status_dispatch.dispatch(LoginStatusAction::LoggedOut);
+        });
+    });
+    html! {
+        <button class={"button is-danger"} onclick={logout_click}>
+            {"Log out"}
+        </button>
+    }
 }
