+++ frontend/src/main.rs
@@ -1,5 +1,5 @@
-use linkdoku_common::{LoginFlowResult, LoginFlowStart};
-use reqwest::{StatusCode, Url};
+use linkdoku_common::LoginFlowResult;
+use reqwest::Url;
 use serde::Deserialize;
 use yew::prelude::*;
 use yew::{function_component, html};
@@ -9,7 +9,6 @@ mod components;
 
 use components::core::{BaseURIProvider, Footer, Navbar};
 use components::login::{LoginStatus, UserProvider};
-use components::user::Avatar;
 
 use crate::components::core::use_api_url;
 use crate::components::login::{LoginStatusAction, LoginStatusDispatcher};
@@ -153,67 +152,6 @@ fn switch(route: &Route) -> Html {
     }
 }
 
-#[function_component(LoginButton)]
-fn login_button() -> Html {
-    let history = use_history().unwrap();
-    let start_google = use_api_url("/login/start/google");
-    let login_click = Callback::from(move |_| {
-        // User clicked login, so we need to redirect the user to the login flow
-        // startup
-        let history = history.clone();
-        let start_google = start_google.clone();
-        wasm_bindgen_futures::spawn_local(async move {
-            let response = reqwest::get(start_google)
-                .await
-                .expect("Unable to make call");
-            if response.status() != StatusCode::OK {
-                // some kind of error
-            } else {
-                let res: LoginFlowStart = response.json().await.unwrap();
-                match res {
-                    LoginFlowStart::Idle => history.push(Route::Root),
-                    LoginFlowStart::Redirect(url) => {
-                        gloo::utils::window().location().set_href(&url).unwrap();
-                    }
-                    LoginFlowStart::Error(err) => {
-                        gloo::console::log!("Failure doing login? {}", err);
-                        history.push(Route::Root);
-                    }
-                }
-            }
-        });
-    });
-
-    html! {
-        <button onclick={login_click}>
-            {"Start login"}
-        </button>
-    }
-}
-
-#[function_component(LogoutButton)]
-fn logout_button() -> Html {
-    let login_status_dispatch =
-        use_context::<LoginStatusDispatcher>().expect("Cannot get login status dispatcher");
-    let history = use_history().unwrap();
-    let clear_login = use_api_url("/login/clear");
-    let logout_click = Callback::from(move |_| {
-        let history = history.clone();
-        let login_status_dispatch = login_status_dispatch.clone();
-        let clear_login = clear_login.clone();
-        wasm_bindgen_futures::spawn_local(async move {
-            reqwest::get(clear_login).await.unwrap();
-            history.push(Route::Root);
-            login_status_dispatch.dispatch(LoginStatusAction::LoggedOut);
-        });
-    });
-    html! {
-        <button onclick={logout_click}>
-            {"Log out"}
-        </button>
-    }
-}
-
 #[function_component(LoginStateShow)]
 fn show_login_state() -> Html {
     let login_status = use_context::<LoginStatus>().expect("Cannot get login status");
@@ -222,8 +160,6 @@ fn show_login_state() -> Html {
         LoginStatus::LoggedOut => {
             html! {
                 <div> {"You are not logged in!"}
-                <br />
-                <LoginButton />
                 </div>
             }
         }
@@ -242,9 +178,6 @@ fn show_login_state() -> Html {
                     } else {
                         html! {}
                     }}
-                    <br />
-                    <Avatar name={name.clone()} email={email} />
-                    <LogoutButton />
                 </div>
             }
         }
